% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{download_dependencies}
\alias{download_dependencies}
\title{Download the dependencies of a package}
\usage{
download_dependencies(dependencies, directory, replace = FALSE)
}
\arguments{
\item{dependencies}{A dataframe. The dependencies of a package.}

\item{directory}{A string (default = "network"). The directory where
downloaded packages are to be stored. Creates the directory if it does not
exist.}

\item{replace}{A boolean (default = FALSE). Control if existing packages in
the directory should be replaced.}
}
\value{
A two-column matrix of names and destination file names of those
  packages successfully downloaded.
}
\description{
Download the dependencies of a package.
}
\examples{
\dontrun{
dependencies <- get_dependencies(package = "shiny")
download_dependencies(dependencies = dependencies, directory = "packages")
}
}
\seealso{
\code{\link{get_dependencies}}, \code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
