\name{flattenPkgDesc}
\alias{flattenPkgDesc}
\title{Function to extract a consistent set of variables from a
\code{\link{packageDescription}} object in a consistent form.}
\usage{
flattenPkgDesc(x)
}
\arguments{
  \item{x}{an object returned by
  \code{\link{packageDescription}}.}
}
\value{
a character vector
}
\description{
Function to extract a consistent set of variables from a
\code{\link{packageDescription}} object in a consistent
form.
}
\author{
Gavin L. Simpson
}

