% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_dependencies}
\alias{get_dependencies}
\title{Get the dependencies of a package}
\usage{
get_dependencies(package, suggest = FALSE, base_packages = FALSE)
}
\arguments{
\item{package}{A string. The package for which you want to retrieve the
dependencies.}

\item{suggest}{A boolean (default = FALSE). If TRUE, retrieves `Suggests`
dependencies (non-recursively).}

\item{base_packages}{A boolean (default = FALSE). If TRUE, includes base R
packages in the dependencies.}
}
\value{
A dataframe with each dependency and the type for each one of them.
}
\description{
Get the dependencies of a package.
}
\examples{
\dontrun{
get_dependencies(package = "shiny")
}
}
\seealso{
\code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
