% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_nodes}
\alias{get_nodes}
\title{Get the nodes from the dependencies}
\usage{
get_nodes(dependencies)
}
\arguments{
\item{dependencies}{A dataframe. The dependencies of a package.}
}
\value{
A vector of the nodes from a dataframe of dependencies.
}
\description{
Get the nodes from the dependencies.
}
\examples{
\dontrun{
dependencies <- get_dependencies(package = "shiny")
get_nodes(dependencies = dependencies)
}
}
\seealso{
\code{\link{get_dependencies}}, \code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
