% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{get_version}
\alias{get_version}
\title{Get the current version on CRAN for a vector of packages}
\usage{
get_version(packages)
}
\arguments{
\item{packages}{A vector of packages.}
}
\value{
A named vector of packages where the value is the current version on
  CRAN.
}
\description{
Get the current version on CRAN for a vector of packages.
}
\examples{
\dontrun{
get_version(packages = c("shiny", "igraph"))
}
}
\seealso{
\code{\link{get_dependencies}}, \code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
