% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{install_dependencies}
\alias{install_dependencies}
\title{Generate the code to install the dependencies of a package}
\usage{
install_dependencies(dependencies, installed = TRUE)
}
\arguments{
\item{dependencies}{A dataframe. The dependencies of a package.}

\item{installed}{A boolean or a character vector (default = TRUE). Filter out
the packages that are already installed.}
}
\value{
The code to install the dependencies of a package.
}
\description{
Generate the code to install the dependencies of a package.
}
\examples{
\dontrun{
dependencies <- get_dependencies(package = "shiny")
install_dependencies(dependencies = dependencies)
}
}
\seealso{
\code{\link{get_dependencies}}, \code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
