% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{plot_dependencies}
\alias{plot_dependencies}
\title{Plot a graph or a tree of the dependencies}
\usage{
plot_dependencies(
  dependencies,
  type = c("network", "tree"),
  colors = NULL,
  ...
)
}
\arguments{
\item{dependencies}{A dataframe. The dependencies of a package.}

\item{type}{A string (default = "network"). Control the type of the output:
"network" or "tree".}

\item{colors}{A vector (default = NULL). Control the colors for each type of
dependency.}

\item{...}{Other arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot or a tree of the dependencies.
}
\description{
Plot a graph or a tree of the dependencies.
}
\examples{
\dontrun{
dependencies <- get_dependencies(package = "shiny")
plot_dependencies(dependencies = dependencies, type = "network")
plot_dependencies(dependencies = dependencies, type = "tree")
}
}
\seealso{
\code{\link{get_dependencies}}, \code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
