% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{plot_graph}
\alias{plot_graph}
\title{Plot a graph of the dependencies}
\usage{
plot_graph(dependencies, colors = NULL, ...)
}
\arguments{
\item{dependencies}{A dataframe. The dependencies of a package.}

\item{colors}{A vector (default = NULL). Control the colors for each type of
dependency.}

\item{...}{Other arguments passed to \code{\link[graphics]{plot}}.}
}
\value{
A plot of the dependencies.
}
\description{
Plot a graph of the dependencies.
}
\examples{
\dontrun{
dependencies <- get_dependencies(package = "shiny")
plot_graph(dependencies = dependencies)
}
}
\seealso{
\code{\link{get_dependencies}}, \code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
