% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dependencies.R
\name{download_dependencies}
\alias{download_dependencies}
\title{Download the dependencies of a package}
\usage{
download_dependencies(
  dependencies,
  directory,
  repository = getOption("repos"),
  installed = TRUE,
  replace = FALSE
)
}
\arguments{
\item{dependencies}{A dataframe. The dependencies of a package.}

\item{directory}{A string (default = "network"). The directory where
the downloaded packages are to be stored. Creates the directory if it does
not exist.}

\item{repository}{A string (default = getOption("repos")).
The base URL of the repository to use.}

\item{installed}{A boolean or a character vector (default = TRUE). Filter out
the packages that are already installed.}

\item{replace}{A boolean (default = FALSE). Control if the existing packages
in the directory should be replaced.}
}
\value{
A two-column matrix of names and destination file names for
  the packages successfully downloaded.
}
\description{
Download the dependencies of a package.
}
\examples{
\dontrun{
dependencies <- get_dependencies(package = "shiny")
download_dependencies(dependencies = dependencies, directory = "packages")
}
}
\seealso{
\code{\link{get_dependencies}}, \code{\link{plot_dependencies}},
  \code{\link{download_dependencies}}, \code{\link{install_dependencies}}
}
